fix: Download and install terraform from the hasicorp.com release page instead of using `go install`
Any version of `terraform` after version 1.9 will no longer install using
`go install` due to error:

```
➜  ~ go install github.com/hashicorp/terraform@latest
go: github.com/hashicorp/terraform@v1.9.2 requires go >= 1.22.4; switching to go1.22.5
go: github.com/hashicorp/terraform@latest (in github.com/hashicorp/terraform@v1.9.2):
        The go.mod file for the module providing named packages contains one or
        more replace directives. It must not contain directives that would cause
        it to be interpreted differently than if it were the main module.
```

Instead we can install directly from the hashicorp releases page.

The release page provides no `current` or `latest` symlink so the best proxy I found was
to use the latest tag in their github releases and download that version from their
release page. This is a zip archive of the `terraform` binary so we need to download unzip and
ensure the permissions are correct.

The `unzip` command was used for this so it too has been added as a depenendency.
For completeness it was added to Wolfi and MacOS but it is only used for Linux installs.

The resulting `terraform` binary is saved in `~/bin/` just like `chainctl`

Signed-off-by: philroche <phil.roche@chainguard.dev>